pyproject.toml
# Dependencies(Using uv)
[project]
name = "weather-energy-pipeline"
version = "1.0.0"
dependencies = [
    "requests",
    "pandas",
    "pyyaml",
    "streamlit",
    "plotly",
    "scikit-learn",
    "numpy"
]